moved items and docs to new files
